+++ crates/miden-standards/src/note/costs/mod.rs
@@ -35,6 +35,7 @@ use crate::note::{
     BurnNote,
     FaucetPolicyActionNote,
     FeeSponsorshipNote,
+    FungibleFaucetConfigNote,
     MintNote,
     NetworkAccountConfigNote,
     OwnerActionNote,
@@ -120,6 +121,9 @@ impl StandardNote {
             StandardNote::MINT => Some(NoteCost::of::<MintNote>()),
             StandardNote::BURN => Some(NoteCost::of::<BurnNote>()),
             StandardNote::FAUCET_POLICY_ACTION => Some(NoteCost::of::<FaucetPolicyActionNote>()),
+            StandardNote::FUNGIBLE_FAUCET_CONFIG => {
+                Some(NoteCost::of::<FungibleFaucetConfigNote>())
+            },
             StandardNote::ALLOWLIST_CONFIG => Some(NoteCost::of::<AllowlistConfigNote>()),
             StandardNote::BLOCKLIST_CONFIG => Some(NoteCost::of::<BlocklistConfigNote>()),
             StandardNote::PAUSE_ACTION => Some(NoteCost::of::<PauseActionNote>()),
@@ -155,6 +159,10 @@ mod tests {
             (MintNote::script_root(), MINT_CONSUMPTION_CYCLES),
             (BurnNote::script_root(), BURN_CONSUMPTION_CYCLES),
             (FaucetPolicyActionNote::script_root(), FAUCET_POLICY_ACTION_CONSUMPTION_CYCLES),
+            (
+                FungibleFaucetConfigNote::script_root(),
+                FUNGIBLE_FAUCET_CONFIG_CONSUMPTION_CYCLES,
+            ),
             (AllowlistConfigNote::script_root(), ALLOWLIST_CONFIG_CONSUMPTION_CYCLES),
             (BlocklistConfigNote::script_root(), BLOCKLIST_CONFIG_CONSUMPTION_CYCLES),
             (PauseActionNote::script_root(), PAUSE_ACTION_CONSUMPTION_CYCLES),